refactor: add subcommand trigger to existing streamson-bin options
